% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.anovapowersim_achieved_power}
\alias{print.anovapowersim_achieved_power}
\title{Print a fixed-sample achieved-power result}
\usage{
\method{print}{anovapowersim_achieved_power}(x, ...)
}
\arguments{
\item{x}{An \code{anovapowersim_achieved_power} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fixed-sample achieved-power result
}
