% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_unbalanced.R
\name{print.anovapowersim_unbalanced_power}
\alias{print.anovapowersim_unbalanced_power}
\title{Print simulated power for an unbalanced design}
\usage{
\method{print}{anovapowersim_unbalanced_power}(x, ...)
}
\arguments{
\item{x}{An \code{anovapowersim_unbalanced_power} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print simulated power for an unbalanced design
}
